- Make actorScramblerMod unbounded; will sum all values to a final result. - Fix glitch with No Visual Lock tooltip - Change save/load mechanic to only occur for saves with combat data - Always show evasion pips in the targeting HUD

+++ LowVisibility/LowVisibility/Patch/CombatHUDPatches.cs
@@ -195,7 +195,7 @@ namespace LowVisibility.Patch {
                 }
 
                 if (lockState.visionLockLevel == VisionLockType.None) {
-                    AddToolTipDetailMethod.GetValue(new object[] { "NO VISUAL LOCK", LowVisibility.Config.SensorsOnlyPenalty });
+                    AddToolTipDetailMethod.GetValue(new object[] { "NO VISUAL LOCK", (int)LowVisibility.Config.SensorsOnlyPenalty });
                 }
 
                 StaticEWState targetEWConfig = State.GetStaticState(target as AbstractActor);
